!! Impose a cap of frequency of positions stored for a group!!
When a position is being stored, quantize it to a fixed interval
(sub-second resolution). This is an easy way to keep a handle on the
amount of position data, so we don't choke on it on playback. I don't
think we don't need to save at the full rate that the input stream is
capable of generating. We still interpolate between positions on
playback, so not much will be lost, is the hope.