Recursive-Details-View-0.1.919
This version has a hack in place for trying out a form of recursive
group animations. If you click the ellipsis button on a character, I
just push a new controller scoped to only that character. Of course,
we'd want this to look more understandable if we want to go this route.

Keyframing is a bit weird in this build (I was fighting between two
different approaches), so stick with recording motions.