Allow for recording scale or rotate over top of existing paths
This required changing the way that we clear positions in advance of
recording. Instead of just nuking all positions after the start time of
the recording, we have to a) stop rendering the motions for the
channels that are being overwritten and b) incrementally 'flatten' the
positions between each new incoming datapoint.