Big Change to keyframes! (maintain for each type of motion)
For each keyframe, track whether it is for S,R,orT (or some
combination). This will let us know whether to clear them or not when
recording over an existing section. Storing the keyframe flag as an
int, with bitwise helper functions for checking which of the flags it
applies to.