Add an erase button
Full disclosure: this eraser is a bit crummy.
It goes through every line and looks for points within a radius of the
touch. If there are hits at the start or end of the line, the line is
truncated. If there are hits in the middle of the line, the line is
split into two lines.
There is no nice inter-point interpolation going on, or fix-up of the
new ends of the line. Results are pretty jagged, but it erases.