Create the selection line without a managed object context
Each object in CoreData is supposed to belong to a managedobjectcontext
(basically like the DB it lives in). That means it becomes part of our
undo/redo stack and on-disk saved state. Since the selection line is
transient, we want to avoid the complication of cleaning up all of that
state. Instead of creating it like all the other lines, we'll just
create it without a managedObjectContext, so it's like a free-floating
object.