Hook up Undo/Redo for line drawing
This encouraged a change to a better model of line drawing. Instead of
putting new lines in the object hierarchy on creation, we make a
temporary line which only gets inserted into the hierarchy when drawing
is completed and we know whether to take it seriously. This has the
side-effect of making automatic undo-redo understand the line properly,
since its creation is a single event.