Fix undo for eraser actions
Undoing erasing was broken. Usually the actions to undo are grouped
together if they happen on the same iteration of the run loop. Our
erasing is actually doing pretty complex stuff to the data structure
(changing lines, adding and deleting groups, etc), so the default
behavior isn't good enough. Instead, when we start erasing, we make a
call to "beginUndoGrouping" and make sure we pair it with
"endUndoGrouping". Everything between those two calls gets treated as a
single undo event!